fix: finalize plugin setup with server url and deployment config
