+++ Procfile
@@ -1,1 +1,1 @@
-web: bash start.sh
+web: sh -c 'uvicorn exo_fin_gpt.exo_interface_api.exo_interface_api:app --host=0.0.0.0 --port=$PORT'
