+++ start.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-uvicorn main:app --host 0.0.0.0 --port ${PORT:-8000}
+uvicorn exo_fin_gpt.exo_interface_api.exo_interface_api:app --host 0.0.0.0 --port ${PORT:-8000}
